Use p4merge
`kdiff` was annoyingly unconfigurable. I was trying to figure out how to
get it to not remove the blank line at the end of the file and it seemed
awkward/impossible to do it without some hacks.

I've decided to move to `p4merge` instead which is currently installable
with:

```
brew install p4v
```